\name{readpresentation-package}
\alias{readpresentation-package}
\alias{readpresentation}
\docType{package}
\title{
  Read in presentation files
}
\description{
  functions created in context of a introductory R course March 2015 to read in presentation file. These functions does not work with presentation files in general.
}
\details{
\tabular{ll}{
Package: \tab readpresentation\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-03-15\cr
License: \tab MIT License\cr
}

}
\author{
Mandy Vogel
Mandy Vogel <mandy.vogel@googlemail.com>
}
\references{
}
\keyword{ package }
\seealso{
}
\examples{
}
